+++ libs/dot-components/src/lib/components/pill/Pill.tsx
@@ -12,8 +12,12 @@ export type PillStatus =
   | 'default';
 
 export interface PillProps extends CommonProps {
+  /** Background color for the pill, ignored if 'status' is not 'default' */
+  backgroundColor?: string;
   /** The text for the pill. */
   label: string;
+  /** Label color for the pill, ignored if 'status' is not 'default' */
+  labelColor?: string;
   /** Determines the size of the pill 'medium' or 'small' */
   size?: PillSize;
   /** Determines the status of the pill component */
@@ -21,9 +25,11 @@ export interface PillProps extends CommonProps {
 }
 
 export const DotPill = ({
-  label,
+  backgroundColor,
   className,
   'data-testid': dataTestId,
+  label,
+  labelColor,
   size = 'medium',
   status = 'default',
 }: PillProps) => {
@@ -31,13 +37,13 @@ export const DotPill = ({
 
   return (
     <StyledPill
+      backgroundColor={backgroundColor}
       classes={{ root: rootClasses }}
-      color="default"
-      data-testid={dataTestId}
       clickable={false}
+      data-testid={dataTestId}
       label={label}
+      labelColor={labelColor}
       size={size}
-      variant="default"
     />
   );
 };